requirements.txt
fastapi
uvicorn[standard]
requests
SQLAlchemy
psycopg2-binary
python-dotenv

